NASA_space_app/requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.9.0
geopandas==1.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.77
langchain-text-splitters==0.3.11
langsmith==0.4.31
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
rasterio==1.4.3
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.16.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zstandard==0.25.0
